@@ -1,15 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 30
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 31
     defaultConfig {
         applicationId "kr.azazel.barcode"
-        minSdkVersion 15
+        minSdkVersion 21
         targetSdkVersion 30
-        versionCode 20
-        versionName "2.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionCode 22
+        versionName "2.2"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -29,22 +28,41 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation project(':framework')
+    implementation 'androidx.core:core:1.7.0'
+    implementation "androidx.cardview:cardview:1.0.0"
     implementation project(':tedbottompicker')
 
-    implementation 'com.android.support:design:24.2.1'
-    implementation 'com.android.support:recyclerview-v7:24.2.1'
-    implementation 'com.android.support:cardview-v7:24.2.1'
+    compileOnly 'org.projectlombok:lombok:1.18.24'
+    annotationProcessor 'org.projectlombok:lombok:1.18.24'
+
+    implementation 'commons-io:commons-io:2.4'
+
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'com.google.android.gms:play-services-vision:9.6.1'
 
     implementation 'com.jakewharton:butterknife:7.0.1'
     implementation 'com.squareup.picasso:picasso:2.5.2'
 
-    implementation 'com.journeyapps:zxing-android-embedded:3.4.0'
+    implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
 
     implementation 'com.theartofdev.edmodo:android-image-cropper:2.3.+'
     implementation 'com.github.rey5137:material:1.2.4'
 
-    implementation 'com.google.firebase:firebase-ads:9.6.1'
-    implementation 'com.google.firebase:firebase-core:9.6.1'
+//    implementation 'com.google.firebase:firebase-ads:9.6.1'
+//    implementation 'com.google.firebase:firebase-core:9.6.1'
+//    // Import the Firebase BoM
+//    implementation platform('com.google.firebase:firebase-bom:30.0.1')
+//
+//    // When using the BoM, you don't specify versions in Firebase library dependencies
+//
+//    // Declare the dependency for the Firebase SDK for Google Analytics
+//    implementation 'com.google.firebase:firebase-analytics'
+//    implementation 'com.google.firebase:firebase-ads'
+//    implementation 'com.google.firebase:firebase-core'
+    implementation 'com.google.android.gms:play-services-ads:20.6.0'
+
+//    implementation 'gun0912.ted:tedbottompicker:2.0.1'
 }
 apply plugin: 'com.google.gms.google-services'
