@@ -4,9 +4,9 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
 
-import com.azazel.framework.util.LOG;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
+import com.azazel.framework.util.LOG;
 
 import java.lang.ref.WeakReference;
 
@@ -48,17 +48,17 @@ abstract public class AzAppCompatActivity extends AppCompatActivity implements I
 
     protected final Handler activityHandler = new InnerClassHandler(this);
 
-    static class InnerClassHandler extends Handler{
+    static class InnerClassHandler extends Handler {
         private final WeakReference<IAzContext> azContextRef;
 
-        InnerClassHandler(IAzContext azContext){
+        InnerClassHandler(IAzContext azContext) {
             this.azContextRef = new WeakReference<IAzContext>(azContext);
         }
 
         @Override
-        public void handleMessage(Message msg){
+        public void handleMessage(Message msg) {
             IAzContext context = azContextRef.get();
-            if(context != null)
+            if (context != null)
                 context.onEventReceived(msg);
         }
     }
