@@ -6,7 +6,7 @@ import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 
 import com.azazel.cafecrawler.data.CrawlDataHelper;
 import com.azazel.cafecrawler.data.CrawlDataHelper.Search;
