@@ -1,8 +1,8 @@
 package com.azazel.cafecrawler;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.core.app.Fragment;
+import androidx.core.app.FragmentManager;
+import androidx.core.app.FragmentPagerAdapter;
 
 import com.azazel.cafecrawler.fragment.ICrawlFragment;
 import com.azazel.framework.view.SmartFragmentStatePagerAdapter;
