@@ -2,24 +2,18 @@ package kr.azazel.barcode;
 
 import android.app.Activity;
 import android.app.Dialog;
-import android.content.DialogInterface;
 import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.media.MediaScannerConnection;
 import android.net.Uri;
-import android.provider.MediaStore;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.azazel.framework.util.AzUtil;
 import com.azazel.framework.util.LOG;
@@ -31,9 +25,6 @@ import com.squareup.picasso.Picasso;
 import com.theartofdev.edmodo.cropper.CropImageView;
 
 import java.io.File;
-import java.io.FileDescriptor;
-import java.io.IOException;
-import java.text.SimpleDateFormat;
 
 import gun0912.tedbottompicker.TedBottomPicker;
 import kr.azazel.barcode.vo.MyBarcode;
@@ -90,8 +81,8 @@ public class PopupUtil {
                     cropImageView.setOnCropImageCompleteListener(new CropImageView.OnCropImageCompleteListener() {
                         @Override
                         public void onCropImageComplete(CropImageView view, CropImageView.CropResult result) {
-                            LOG.d(TAG, "onCropImageComplete - " + (result==null?"null":result.isSuccessful()));
-                            if(result != null && result.isSuccessful()) {
+                            LOG.d(TAG, "onCropImageComplete - " + (result == null ? "null" : result.isSuccessful()));
+                            if (result != null && result.isSuccessful()) {
                                 dialog.cancel();
                                 listener.onCropImageComplete(view, result);
                             }
@@ -171,7 +162,7 @@ public class PopupUtil {
                             @Override
                             public void onImageSelected(Uri uri) {
                                 LOG.d(TAG, "onImageSelected - uri : " + uri);
-                                if(uri != null) {
+                                if (uri != null) {
                                     showCoverImageCropPopup(activity, uri, new CropImageView.OnCropImageCompleteListener() {
                                         @Override
                                         public void onCropImageComplete(CropImageView view, CropImageView.CropResult result) {
@@ -182,7 +173,7 @@ public class PopupUtil {
                                             }
                                         }
                                     });
-                                }else{
+                                } else {
 
                                 }
                             }
@@ -193,8 +184,8 @@ public class PopupUtil {
                 View.OnClickListener clickListener = new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        switch (v.getId()){
-                            case R.id.btn_ok:{
+                        switch (v.getId()) {
+                            case R.id.btn_ok: {
                                 boolean saved = MyBarcode.saveBarcode(selectedCategory.value(), code.rawValue, etTitle.getText().toString(), code.format
                                         , etDesc.getText().toString(), etBrand.getText().toString(), org, codeImg, bitmapCover, expirationDt);
 
@@ -230,26 +221,26 @@ public class PopupUtil {
                                 dialog.cancel();
                                 break;
                             }
-                            case R.id.btn_cancel:{
+                            case R.id.btn_cancel: {
                                 dialog.cancel();
                                 break;
                             }
-                            case R.id.img_cover:{
-                                imagePicker.show(((AppCompatActivity)activity).getSupportFragmentManager());
+                            case R.id.img_cover: {
+                                imagePicker.show(((AppCompatActivity) activity).getSupportFragmentManager());
                                 break;
                             }
-                            case R.id.tv_expiredt_reset:{
+                            case R.id.tv_expiredt_reset: {
                                 expirationDt = 0;
                                 tvExpirationDt.setText(R.string.no_expiredt);
                                 break;
                             }
                             case R.id.tv_expiredt_value:
-                            case R.id.tv_expiredt:{
+                            case R.id.tv_expiredt: {
                                 boolean isLightTheme = ThemeManager.getInstance().getCurrentTheme() == 0;
-                                DatePickerDialog.Builder builder = new DatePickerDialog.Builder(isLightTheme ? R.style.Material_App_Dialog_DatePicker_Light :  R.style.Material_App_Dialog_DatePicker){
+                                DatePickerDialog.Builder builder = new DatePickerDialog.Builder(isLightTheme ? R.style.Material_App_Dialog_DatePicker_Light : R.style.Material_App_Dialog_DatePicker) {
                                     @Override
                                     public void onPositiveActionClicked(DialogFragment fragment) {
-                                        DatePickerDialog dateDialog = (DatePickerDialog)fragment.getDialog();
+                                        DatePickerDialog dateDialog = (DatePickerDialog) fragment.getDialog();
                                         expirationDt = dateDialog.getDate();
 
                                         tvExpirationDt.setText(AzUtil.getLongDateStringFromMils(dialog.getContext(), expirationDt, false));
@@ -263,7 +254,7 @@ public class PopupUtil {
                                     }
                                 };
 
-                                if(expirationDt > 0)
+                                if (expirationDt > 0)
                                     builder.date(expirationDt);
                                 builder.positiveAction(activity.getString(android.R.string.ok))
                                         .negativeAction(activity.getString(android.R.string.cancel));
@@ -344,9 +335,9 @@ public class PopupUtil {
                 final Uri imageCode = Uri.fromFile(new File(barcode.barcodeImage));
                 Picasso.with(activity).load(imageCode).into(imgCode);
 
-                if(TextUtils.isEmpty(barcode.coverImage)) {
+                if (TextUtils.isEmpty(barcode.coverImage)) {
                     imgCover.setImageResource(R.mipmap.ic_default);
-                }else{
+                } else {
                     final Uri imageCover = Uri.fromFile(new File(barcode.coverImage));
                     Picasso.with(activity).load(imageCover).into(imgCover);
                 }
@@ -360,7 +351,7 @@ public class PopupUtil {
                             @Override
                             public void onCropImageComplete(CropImageView view, CropImageView.CropResult result) {
                                 LOG.d(TAG, "onCropImageComplete - result : " + result);
-                                if(result != null) {
+                                if (result != null) {
                                     bitmapCover = result.getBitmap();
                                     imgCover.setImageBitmap(bitmapCover);
                                 }
@@ -377,8 +368,8 @@ public class PopupUtil {
                 View.OnClickListener clickListener = new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        switch (v.getId()){
-                            case R.id.btn_ok:{
+                        switch (v.getId()) {
+                            case R.id.btn_ok: {
                                 barcode.title = etTitle.getText().toString();
                                 barcode.description = etDesc.getText().toString();
                                 barcode.category = selectedCategory.value();
@@ -387,26 +378,26 @@ public class PopupUtil {
                                 dialog.cancel();
                                 break;
                             }
-                            case R.id.btn_close:{
+                            case R.id.btn_close: {
                                 dialog.cancel();
                                 break;
                             }
-                            case R.id.img_cover:{
-                                imagePicker.show(((AppCompatActivity)activity).getSupportFragmentManager());
+                            case R.id.img_cover: {
+                                imagePicker.show(((AppCompatActivity) activity).getSupportFragmentManager());
                                 break;
                             }
-                            case R.id.tv_expiredt_reset:{
+                            case R.id.tv_expiredt_reset: {
                                 barcode.expirationDate = 0;
                                 tvExpirationDt.setText(R.string.no_expiredt);
                                 break;
                             }
                             case R.id.tv_expiredt_value:
-                            case R.id.tv_expiredt:{
+                            case R.id.tv_expiredt: {
                                 boolean isLightTheme = ThemeManager.getInstance().getCurrentTheme() == 0;
-                                DatePickerDialog.Builder builder = new DatePickerDialog.Builder(isLightTheme ? R.style.Material_App_Dialog_DatePicker_Light :  R.style.Material_App_Dialog_DatePicker){
+                                DatePickerDialog.Builder builder = new DatePickerDialog.Builder(isLightTheme ? R.style.Material_App_Dialog_DatePicker_Light : R.style.Material_App_Dialog_DatePicker) {
                                     @Override
                                     public void onPositiveActionClicked(DialogFragment fragment) {
-                                        DatePickerDialog dateDialog = (DatePickerDialog)fragment.getDialog();
+                                        DatePickerDialog dateDialog = (DatePickerDialog) fragment.getDialog();
                                         barcode.expirationDate = dateDialog.getDate();
 
                                         tvExpirationDt.setText(AzUtil.getLongDateStringFromMils(dialog.getContext(), barcode.expirationDate, false));
@@ -420,7 +411,7 @@ public class PopupUtil {
                                     }
                                 };
 
-                                if(barcode.expirationDate > 0)
+                                if (barcode.expirationDate > 0)
                                     builder.date(barcode.expirationDate);
                                 builder.positiveAction(activity.getString(android.R.string.ok))
                                         .negativeAction(activity.getString(android.R.string.cancel));
@@ -443,7 +434,7 @@ public class PopupUtil {
                 dialog.findViewById(R.id.tv_expiredt_value).setOnClickListener(clickListener);
                 dialog.findViewById(R.id.tv_expiredt_reset).setOnClickListener(clickListener);
 
-                if(barcode.expirationDate > 0)
+                if (barcode.expirationDate > 0)
                     tvExpirationDt.setText(AzUtil.getLongDateStringFromMils(dialog.getContext(), barcode.expirationDate, false));
             }
         }).show();
