@@ -1,10 +1,12 @@
 package com.azazel.framework.view;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+
 import android.util.SparseArray;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 /* 
 Extension of FragmentStatePagerAdapter which intelligently caches 
@@ -12,30 +14,30 @@ all active fragments and manages the fragment lifecycles.
 Usage involves extending from SmartFragmentStatePagerAdapter as you would any other PagerAdapter.
 */
 public abstract class SmartFragmentStatePagerAdapter extends FragmentStatePagerAdapter {
-	// Sparse array to keep track of registered fragments in memory
-	private SparseArray<Fragment> registeredFragments = new SparseArray<Fragment>();
+    // Sparse array to keep track of registered fragments in memory
+    private SparseArray<Fragment> registeredFragments = new SparseArray<Fragment>();
 
-	public SmartFragmentStatePagerAdapter(FragmentManager fragmentManager) {
-		super(fragmentManager);
-	}
+    public SmartFragmentStatePagerAdapter(FragmentManager fragmentManager) {
+        super(fragmentManager);
+    }
 
-	// Register the fragment when the item is instantiated
-	@Override
-	public Object instantiateItem(ViewGroup container, int position) {
-		Fragment fragment = (Fragment) super.instantiateItem(container, position);
-		registeredFragments.put(position, fragment);
-		return fragment;
-	}
+    // Register the fragment when the item is instantiated
+    @Override
+    public Object instantiateItem(ViewGroup container, int position) {
+        Fragment fragment = (Fragment) super.instantiateItem(container, position);
+        registeredFragments.put(position, fragment);
+        return fragment;
+    }
 
-	// Unregister when the item is inactive
-	@Override
-	public void destroyItem(ViewGroup container, int position, Object object) {
-		registeredFragments.remove(position);
-		super.destroyItem(container, position, object);
-	}
+    // Unregister when the item is inactive
+    @Override
+    public void destroyItem(ViewGroup container, int position, Object object) {
+        registeredFragments.remove(position);
+        super.destroyItem(container, position, object);
+    }
 
-	// Returns the fragment for the position (if instantiated)
-	public Fragment getRegisteredFragment(int position) {
-		return registeredFragments.get(position);
-	}
+    // Returns the fragment for the position (if instantiated)
+    public Fragment getRegisteredFragment(int position) {
+        return registeredFragments.get(position);
+    }
 }
